refactor(stock): stock 도메인 구조 리팩토링
 Stock API 라우트를 app/api/stocks로 이동
 Stock 페이지들을 /src/app에서 /app 디렉토리로 이동 (searchstocktype, selectstocktype)
 Stock API 함수 정리 및 불필요한 함수 제거 (getSelectedStocks)
 Stock 쿼리 훅 정리 (useSelectedStocks, useInfiniteSelectedStocks 제거)
 Stock 액션 개선 (getHotStocks에 market 파라미터 추가)
 Stock 서버 리포지토리 import 경로를 @/app/prisma로 변경
 Stock 선택 페이지 개선 (NAS, NYS 시장 주식 조합 및 정렬)
 Stock UI 컴포넌트를 screens/select-stock으로 이동
 API 경로를 상수로 변경하여 유지보수성 향상
