+++ apps/web/app/api/stocks/[symbol]/route.ts
