+++ apps/web/app/api/stocks/hot/route.ts
