+++ apps/web/app/api/stocks/route.ts
