+++ apps/web/app/api/stocks/select/route.ts
