+++ apps/web/app/selectstocktype/layout.tsx
