+++ apps/web/app/selectstocktype/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/screens/select-stock';
