--- apps/web/src/app/selectstocktype/page.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/_pages/select-stock';
