+++ apps/web/src/features/stock/api/stock.ts
@@ -1,39 +1,33 @@
-import { httpWithCredentials, http } from '@/shared/api/client';
+import { http } from '@/shared/api/client';
 import type { OverseasStockPriceResponse } from '../server';
 import type { Stock, DbStock, MarketCode } from '@/entities/stock';
+import { API_PATH } from '@/shared/config';
 
 export async function getStocks({ count = 30, cursor = 0 }: { count?: number; cursor?: number } = {}) {
-    const res = await http.get<Stock[]>('/api/stocks', { params: { count, cursor } });
+    const res = await http.get<Stock[]>(API_PATH.STOCKS, { params: { count, cursor } });
     return res.data;
 }
 
 export async function selectStock(stockSymbols: string[]) {
-    const res = await httpWithCredentials.post('/api/stocks/select', { stockSymbols });
-    return res.data;
-}
-
-export async function getSelectedStocks({ count = 20, cursor = 0 }: { count?: number; cursor?: number } = {}) {
-    const res = await httpWithCredentials.get<Stock[]>('/api/stocks/selected', {
-        params: { count, cursor },
-    });
+    const res = await http.post(API_PATH.STOCKS_SELECT, { stockSymbols });
     return res.data;
 }
 
 export async function getOverseasStockPrice({ symbol }: { symbol: string }) {
-    const res = await http.get<OverseasStockPriceResponse>(`/api/stocks/price/overseas`, {
+    const res = await http.get<OverseasStockPriceResponse>(API_PATH.STOCKS_PRICE_OVERSEAS, {
         params: { symbol },
     });
     return res.data;
 }
 
 export async function getHotStock({ market }: { market: MarketCode }) {
-    const res = await http.get<Stock[]>('/api/stocks/hot', {
+    const res = await http.get<Stock[]>(API_PATH.STOCKS_HOT, {
         params: { market },
     });
     return res.data;
 }
 
 export async function getStockBySymbol({ symbol }: { symbol: string }) {
-    const res = await http.get<DbStock>(`/api/stocks/${symbol}`);
+    const res = await http.get<DbStock>(`${API_PATH.STOCKS}/${symbol}`);
     return res.data;
 }
