+++ apps/web/src/features/stock/index.ts
@@ -1,11 +1,4 @@
 export { getHotStocks, getSelectedStockSymbols, selectStockAction } from './action/stock';
-export {
-    getStocks,
-    selectStock,
-    getSelectedStocks,
-    getOverseasStockPrice,
-    getHotStock,
-    getStockBySymbol,
-} from './api/stock';
+export { getStocks, selectStock, getOverseasStockPrice, getHotStock, getStockBySymbol } from './api/stock';
 
 export * from './query';
