+++ apps/web/src/screens/select-stock/hooks/index.ts
