+++ apps/web/src/screens/select-stock/hooks/useInfiniteStocks.ts
