+++ apps/web/src/screens/select-stock/index.ts
