+++ apps/web/src/screens/select-stock/ui/SelectStockClientPage.tsx
