+++ apps/web/src/screens/select-stock/ui/StockChip.tsx
