+++ apps/web/src/screens/select-stock/ui/StockList.tsx
